Changed permit percept to return an integer instead of a double

+++ tygron-connect-environment/src/main/java/nl/tudelft/contextproject/tygron/eis/entities/Controller.java
@@ -122,7 +122,7 @@ public class Controller {
    * @return the number of permits of this stakeholder.
    */
   @AsPercept(name = "permits", filter = Filter.Type.ON_CHANGE)
-  public double permits() {
+  public int permits() {
     return env.requestsOpen();
   }
